Split crates/harn-cli/src/cli.rs by subcommand group (#942)
cli.rs had grown to 4,068 LOC of clap definitions plus 1,677 LOC of
parser tests in a single file. Splitting it into per-subcommand
modules under crates/harn-cli/src/cli/ keeps each file focused and
makes future syntax changes (a new flag on `harn orchestrator
serve`, a new subcommand group, etc.) touch only the relevant
module.

The top-level Cli struct and Command enum stay in cli/mod.rs and
re-export per-subcommand types via `pub(crate) use`, so external
callers continue using `crate::cli::TypeName` unchanged. The 61
parser unit tests move to cli/tests.rs verbatim.

Verified that --help is byte-identical for the top-level command
and all 44 subcommands, and that all 61 cli::tests cases pass.
