+++ crates/harn-cli/src/cli/doctor.rs
@@ -1,0 +1,8 @@
+use clap::Args;
+
+#[derive(Debug, Args)]
+pub(crate) struct DoctorArgs {
+    /// Skip provider connectivity checks.
+    #[arg(long)]
+    pub no_network: bool,
+}
