+++ crates/harn-cli/src/cli/portal.rs
@@ -1,0 +1,17 @@
+use clap::{ArgAction, Args};
+
+#[derive(Debug, Args)]
+pub(crate) struct PortalArgs {
+    /// Directory containing persisted run records.
+    #[arg(long, default_value = ".harn-runs")]
+    pub dir: String,
+    /// Host interface to bind.
+    #[arg(long, default_value = "127.0.0.1")]
+    pub host: String,
+    /// Port to serve the portal on.
+    #[arg(long, default_value_t = 4721)]
+    pub port: u16,
+    /// Open the portal in a browser after starting.
+    #[arg(long, default_value_t = true, action = ArgAction::Set)]
+    pub open: bool,
+}
