+++ crates/harn-cli/src/cli/run.rs
@@ -1,0 +1,55 @@
+use clap::Args;
+
+#[derive(Debug, Args)]
+pub(crate) struct RunArgs {
+    /// Print the LLM trace summary after execution.
+    #[arg(long)]
+    pub trace: bool,
+    /// Deny specific builtins as a comma-separated list.
+    #[arg(long, conflicts_with = "allow")]
+    pub deny: Option<String>,
+    /// Allow only the listed builtins as a comma-separated list.
+    #[arg(long, conflicts_with = "deny")]
+    pub allow: Option<String>,
+    /// Evaluate inline Harn code instead of a file.
+    #[arg(short = 'e')]
+    pub eval: Option<String>,
+    /// Extra skill-discovery roots. Repeatable; each path is a
+    /// directory of `<name>/SKILL.md` bundles, equivalent to a
+    /// single-entry `$HARN_SKILLS_PATH`. Highest-priority layer —
+    /// wins ties against every other layer. See `docs/src/skills.md`.
+    #[arg(long = "skill-dir", value_name = "PATH")]
+    pub skill_dir: Vec<String>,
+    /// Replay LLM responses from a JSONL fixture file instead of
+    /// calling the configured provider.
+    #[arg(
+        long = "llm-mock",
+        value_name = "PATH",
+        conflicts_with = "llm_mock_record"
+    )]
+    pub llm_mock: Option<String>,
+    /// Record executed LLM responses into a JSONL fixture file.
+    #[arg(
+        long = "llm-mock-record",
+        value_name = "PATH",
+        conflicts_with = "llm_mock"
+    )]
+    pub llm_mock_record: Option<String>,
+    /// Emit a signed provenance receipt after the run.
+    #[arg(long)]
+    pub attest: bool,
+    /// Write the signed provenance receipt to this path instead of `.harn/receipts/`.
+    #[arg(long = "receipt-out", value_name = "PATH", requires = "attest")]
+    pub receipt_out: Option<String>,
+    /// Agent id used to look up or generate the receipt signing key.
+    #[arg(long = "attest-agent", value_name = "ID", requires = "attest")]
+    pub attest_agent: Option<String>,
+    /// Path to the .harn file to execute.
+    pub file: Option<String>,
+    /// Positional arguments passed to the pipeline as the global `argv`
+    /// list. Place them after a `--` separator: `harn run script.harn -- a b c`.
+    // `last = true` alone routes post-`--` tokens into `argv`; combining it
+    // with `trailing_var_arg = true` panics at clap runtime.
+    #[arg(last = true)]
+    pub argv: Vec<String>,
+}
