+++ crates/harn-cli/src/cli/skill.rs
@@ -1,0 +1,77 @@
+use clap::{Args, Subcommand};
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillArgs {
+    #[command(subcommand)]
+    pub command: SkillCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum SkillCommand {
+    /// Manage Ed25519 signing keys for skill provenance.
+    Key(SkillKeyArgs),
+    /// Sign a skill manifest and emit `<path>.sig`.
+    Sign(SkillSignArgs),
+    /// Verify a skill manifest against the trusted signer set.
+    Verify(SkillVerifyArgs),
+    /// Manage the local trusted signer registry.
+    Trust(SkillTrustArgs),
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillKeyArgs {
+    #[command(subcommand)]
+    pub command: SkillKeyCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum SkillKeyCommand {
+    /// Generate an Ed25519 keypair and write PEM files to disk.
+    Generate(SkillKeyGenerateArgs),
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillKeyGenerateArgs {
+    /// Path for the private-key PEM. The public key is written to `<path>.pub`.
+    #[arg(long, value_name = "PATH")]
+    pub out: String,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillSignArgs {
+    /// Path to the skill manifest to sign (typically `SKILL.md`).
+    pub skill: String,
+    /// Path to the private-key PEM generated by `harn skill key generate`.
+    #[arg(long, value_name = "PATH")]
+    pub key: String,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillVerifyArgs {
+    /// Path to the skill manifest to verify.
+    pub skill: String,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillTrustArgs {
+    #[command(subcommand)]
+    pub command: SkillTrustCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum SkillTrustCommand {
+    /// Import a trusted signer from a PEM file or URL.
+    Add(SkillTrustAddArgs),
+    /// List the trusted signer fingerprints currently installed locally.
+    List(SkillTrustListArgs),
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct SkillTrustAddArgs {
+    /// PEM source for the public key. Accepts a local path or URL.
+    #[arg(long = "from", value_name = "URL|FILE")]
+    pub from: String,
+}
+
+#[derive(Debug, Args, Default)]
+pub(crate) struct SkillTrustListArgs {}
